added utils class and refactored code
