+++ src/app/plan/page.tsx
@@ -1,12 +1,11 @@
-import { getStock, getTrades } from "../lib/stockService";
+import { getStock } from "../lib/stockService";
 import PlanDashboard from "./PlanDashboard";
 
 export default async function PlanPage() {
   const response = await getStock();
-  const trades = await getTrades();
   return (
     <div style={{ background: "#FAFAFA" }}>
-      <PlanDashboard stocks={response} trades={trades} />
+      <PlanDashboard stocks={response} />
     </div>
   );
 }
