+++ src/app/stocks/TotalCard.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Box, Typography } from "@mui/material";
 
 export interface TotalCardProps {
